+++ R/trelloR-package.R
@@ -1,0 +1,11 @@
+#' @description
+#' An R client for the Trello API. Supports free-tier features such as access
+#' to private boards, creating and updating cards and other resources, and downloading
+#' data in a structured way.
+
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
+NULL